move TDAmeritradeWebSocket into its own class to wrap websocket
